implemented #21, made calcDist & pow globalfunc
