+++ src/helperFiles/globalFunctions.c
@@ -14,3 +14,15 @@ float randFloat(float max) {
 float randFloatRange(float min, float max) {
     return min + (float) rand() * (max - min) / (float) RAND_MAX;
 }
+
+float powTwo(float n) {
+    return n * n;
+}
+
+float powThree(float n) {
+    return n * n * n;
+}
+
+float calcDistance(float diffX, float diffY) {
+    return sqrt(powTwo(diffX) + powTwo(diffY));
+}
